requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic[email]==2.7.1
slowapi==0.1.9
httpx==0.27.0
